feat: implement scale safety and amortized O(1) editing
- Upgraded coordinates and buffer lengths to size_t for large file support (>2GB).
- Implemented exponential line capacity growth for amortized O(1) editing.
- Hardened I/O and bounds checking to prevent path truncation and segmentation faults.
- Updated GEMINI.md to reflect modular architecture and technical standards.

+++ include/stitch/buffer/engine.h
@@ -3,8 +3,8 @@
 
 #include "stitch/types.h"
 
-void buffer_insert_line(StitchBuffer *buf, int at, char *s, size_t len);
-void buffer_del_line(StitchBuffer *buf, int at);
+void buffer_insert_line(StitchBuffer *buf, size_t at, char *s, size_t len);
+void buffer_del_line(StitchBuffer *buf, size_t at);
 void buffer_free(StitchBuffer *buf);
 
 void buffer_insert_char(StitchBuffer *buf, StitchView *view, int c);
