+++ src/buffer/buffer_internal.h
@@ -7,8 +7,8 @@
 void buffer_update_line(Line *line);
 void buffer_row_insert_char(StitchBuffer *buf, Line *line, int at, int c);
 void buffer_row_del_char(StitchBuffer *buf, Line *line, int at);
-void buffer_insert_line(StitchBuffer *buf, int at, char *s, size_t len);
-void buffer_del_line(StitchBuffer *buf, int at);
+void buffer_insert_line(StitchBuffer *buf, size_t at, char *s, size_t len);
+void buffer_del_line(StitchBuffer *buf, size_t at);
 void buffer_free_line(Line *line);
 
 #endif
